+++ src/mup-api.js
@@ -30,24 +30,26 @@ export default class MupAPI {
     let invalid = valid.errors.length > 0 || valid.warnings.length > 0;
 
     if (invalid) {
-      console.log(`loaded mup.js from ${configPath}`)
+      console.log(`loaded mup.js from ${configPath}`);
     }
 
     if (valid.errors.length > 0) {
       console.log(`mup.js has ${valid.errors.length} errors:`);
-      valid.errors.forEach((error) => {
+      valid.errors.forEach(error => {
         console.log(`  - ${error}`);
       });
     }
     if (valid.warnings.length > 0) {
       console.log(`mup.js has ${valid.warnings.length} warnings:`);
-      valid.warnings.forEach((warning) => {
+      valid.warnings.forEach(warning => {
         console.log(`   - ${warning}`);
       });
     }
 
     if (invalid) {
-      console.log('If you think there is a bug in the mup.js validator, please');
+      console.log(
+        'If you think there is a bug in the mup.js validator, please'
+      );
       console.log('create an issue at https://github.com/zodern/meteor-up.');
     }
   }
@@ -167,14 +169,23 @@ export default class MupAPI {
       }
 
       if (info.pem) {
-        auth.pem = fs.readFileSync(resolve(info.pem), 'utf8');
+        try {
+          auth.pem = fs.readFileSync(resolve(info.pem), 'utf8');
+        } catch (e) {
+          console.error(`Unable to load pem at "${resolve(info.pem)}"`);
+          console.error(`for server "${name}"`)
+          if (e.code !== 'ENOENT') {
+            console.log(e);
+          }
+          process.exit(1);
+        }
       } else if (info.password) {
         auth.password = info.password;
       } else if (sshAgent && fs.existsSync(sshAgent)) {
         opts.ssh.agent = sshAgent;
       } else {
         console.error(
-          'error: server %s doesn\'t have password, ssh-agent or pem',
+          "error: server %s doesn't have password, ssh-agent or pem",
           name
         );
         process.exit(1);