+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# 1.2.2
 - Configure nginx max client upload size, and increase default to `10M` (@shadowcodex)
 - Displays better message if it can not find the meteor app
 - Displays message if can not find pem for server